# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the Cart-to-WhatsApp Widget platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface SignupBody {
  owner_email: string;
  /** @minLength 8 */
  password: string;
  store_domain: string;
}

export interface LoginBody {
  owner_email: string;
  password: string;
}

export type StoreMode = typeof StoreMode[keyof typeof StoreMode];


export const StoreMode = {
  personalized: 'personalized',
  discount: 'discount',
} as const;

export interface Store {
  id: string;
  owner_email: string;
  store_domain: string;
  whatsapp_number?: string | null;
  sms_number?: string | null;
  mode: StoreMode;
  discount_amount?: number | null;
  currency: string;
  created_at: string;
}

/**
 * Minimal config needed by the widget on load
 */
export interface StoreConfig {
  whatsapp_number?: string | null;
  sms_number?: string | null;
  mode: StoreMode;
  discount_amount?: number | null;
  currency: string;
}

export interface CreateStoreBody {
  owner_email: string;
  store_domain: string;
  whatsapp_number: string;
  sms_number?: string | null;
  mode: StoreMode;
  discount_amount?: number | null;
  currency?: string;
}

export interface UpdateStoreBody {
  whatsapp_number?: string;
  sms_number?: string | null;
  mode?: StoreMode;
  discount_amount?: number | null;
  currency?: string;
}

export type CaptureChannel = typeof CaptureChannel[keyof typeof CaptureChannel];


export const CaptureChannel = {
  whatsapp: 'whatsapp',
  sms: 'sms',
} as const;

/**
 * Cart items and prices at the time of capture
 */
export type CaptureCartSnapshot = { [key: string]: unknown };

export interface Capture {
  id: string;
  store_id: string;
  /**
     * @minimum 1
     * @maximum 2
     */
  tier: number;
  channel: CaptureChannel;
  /** Cart items and prices at the time of capture */
  cart_snapshot: CaptureCartSnapshot;
  created_at: string;
}

/**
 * Cart items and prices at the time of capture
 */
export type CreateCaptureBodyCartSnapshot = { [key: string]: unknown };

export interface CreateCaptureBody {
  store_id: string;
  /**
     * @minimum 1
     * @maximum 2
     */
  tier: number;
  channel: CaptureChannel;
  /** Cart items and prices at the time of capture */
  cart_snapshot: CreateCaptureBodyCartSnapshot;
}

export interface CaptureList {
  captures: Capture[];
  total: number;
}

export type GetStoreCapturesParams = {
format?: GetStoreCapturesFormat;
};

export type GetStoreCapturesFormat = typeof GetStoreCapturesFormat[keyof typeof GetStoreCapturesFormat];


export const GetStoreCapturesFormat = {
  json: 'json',
  csv: 'csv',
} as const;