# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for the Cart-to-WhatsApp Widget platform
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Dashboard authentication
  - name: stores
    description: Store management
  - name: captures
    description: Capture logging
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a dashboard account
      description: Creates a store owner account and its store row in one step. Starts a session and redirects to settings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: Account created and session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in to the dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out of the dashboard
      responses:
        "200":
          description: Session ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the logged-in store owner
      responses:
        "200":
          description: Current store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stores:
    post:
      operationId: createStore
      tags: [stores]
      summary: Create a new store
      description: Creates a store record on signup. Returns the store including its UUID which acts as the install token for the widget snippet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreBody"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stores/{id}/config:
    get:
      operationId: getStoreConfig
      tags: [stores]
      summary: Get widget config for a store
      description: Called by the widget on load to fetch whatsapp_number, mode, discount_amount, and currency instead of using hardcoded JS constants.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Store config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreConfig"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stores/{id}:
    patch:
      operationId: updateStore
      tags: [stores]
      summary: Update store config
      description: Update store settings from the dashboard (whatsapp number, mode, discount amount, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreBody"
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /captures:
    post:
      operationId: createCapture
      tags: [captures]
      summary: Log a widget capture
      description: Widget posts here when a shopper clicks the WhatsApp or SMS link. Records cart snapshot, tier, and channel.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaptureBody"
      responses:
        "201":
          description: Capture logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capture"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stores/{id}/captures:
    get:
      operationId: getStoreCaptures
      tags: [captures]
      summary: List captures for a store
      description: Returns all captures for a store. Pass ?format=csv to get a CSV export instead of JSON.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        "200":
          description: Captures list (JSON or CSV depending on format param)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureList"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    SignupBody:
      type: object
      properties:
        owner_email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        store_domain:
          type: string
      required:
        - owner_email
        - password
        - store_domain

    LoginBody:
      type: object
      properties:
        owner_email:
          type: string
          format: email
        password:
          type: string
      required:
        - owner_email
        - password

    StoreMode:
      type: string
      enum: [personalized, discount]

    Store:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_email:
          type: string
          format: email
        store_domain:
          type: string
        whatsapp_number:
          type: string
          nullable: true
        sms_number:
          type: string
          nullable: true
        mode:
          $ref: "#/components/schemas/StoreMode"
        discount_amount:
          type: integer
          nullable: true
        currency:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - owner_email
        - store_domain
        - mode
        - currency
        - created_at

    StoreConfig:
      type: object
      description: Minimal config needed by the widget on load
      properties:
        whatsapp_number:
          type: string
          nullable: true
        sms_number:
          type: string
          nullable: true
        mode:
          $ref: "#/components/schemas/StoreMode"
        discount_amount:
          type: integer
          nullable: true
        currency:
          type: string
      required:
        - mode
        - currency

    CreateStoreBody:
      type: object
      properties:
        owner_email:
          type: string
          format: email
        store_domain:
          type: string
        whatsapp_number:
          type: string
        sms_number:
          type: string
          nullable: true
        mode:
          $ref: "#/components/schemas/StoreMode"
        discount_amount:
          type: integer
          nullable: true
        currency:
          type: string
          default: "₹"
      required:
        - owner_email
        - store_domain
        - whatsapp_number
        - mode

    UpdateStoreBody:
      type: object
      properties:
        whatsapp_number:
          type: string
        sms_number:
          type: string
          nullable: true
        mode:
          $ref: "#/components/schemas/StoreMode"
        discount_amount:
          type: integer
          nullable: true
        currency:
          type: string

    CaptureChannel:
      type: string
      enum: [whatsapp, sms]

    Capture:
      type: object
      properties:
        id:
          type: string
          format: uuid
        store_id:
          type: string
          format: uuid
        tier:
          type: integer
          minimum: 1
          maximum: 2
        channel:
          $ref: "#/components/schemas/CaptureChannel"
        cart_snapshot:
          type: object
          description: Cart items and prices at the time of capture
        created_at:
          type: string
          format: date-time
      required:
        - id
        - store_id
        - tier
        - channel
        - cart_snapshot
        - created_at

    CreateCaptureBody:
      type: object
      properties:
        store_id:
          type: string
          format: uuid
        tier:
          type: integer
          minimum: 1
          maximum: 2
        channel:
          $ref: "#/components/schemas/CaptureChannel"
        cart_snapshot:
          type: object
          description: Cart items and prices at the time of capture
      required:
        - store_id
        - tier
        - channel
        - cart_snapshot

    CaptureList:
      type: object
      properties:
        captures:
          type: array
          items:
            $ref: "#/components/schemas/Capture"
        total:
          type: integer
      required:
        - captures
        - total