# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the Cart-to-WhatsApp Widget platform
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod/v4';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * Creates a store owner account and its store row in one step. Starts a session and redirects to settings.
 * @summary Create a dashboard account
 */
export const signupBodyPasswordMin = 8;



export const SignupBody = zod.object({
  "owner_email": zod.email(),
  "password": zod.string().min(signupBodyPasswordMin),
  "store_domain": zod.string()
})

export const SignupResponse = zod.object({
  "id": zod.uuid(),
  "owner_email": zod.email(),
  "store_domain": zod.string(),
  "whatsapp_number": zod.string().nullish(),
  "sms_number": zod.string().nullish(),
  "mode": zod.enum(['personalized', 'discount']),
  "discount_amount": zod.int().nullish(),
  "currency": zod.string(),
  "created_at": zod.coerce.date()
})


/**
 * @summary Log in to the dashboard
 */
export const LoginBody = zod.object({
  "owner_email": zod.email(),
  "password": zod.string()
})

export const LoginResponse = zod.object({
  "id": zod.uuid(),
  "owner_email": zod.email(),
  "store_domain": zod.string(),
  "whatsapp_number": zod.string().nullish(),
  "sms_number": zod.string().nullish(),
  "mode": zod.enum(['personalized', 'discount']),
  "discount_amount": zod.int().nullish(),
  "currency": zod.string(),
  "created_at": zod.coerce.date()
})


/**
 * @summary Log out of the dashboard
 */
export const LogoutResponse = zod.object({
  "success": zod.boolean()
})


/**
 * @summary Get the logged-in store owner
 */
export const GetCurrentUserResponse = zod.object({
  "id": zod.uuid(),
  "owner_email": zod.email(),
  "store_domain": zod.string(),
  "whatsapp_number": zod.string().nullish(),
  "sms_number": zod.string().nullish(),
  "mode": zod.enum(['personalized', 'discount']),
  "discount_amount": zod.int().nullish(),
  "currency": zod.string(),
  "created_at": zod.coerce.date()
})


/**
 * Creates a store record on signup. Returns the store including its UUID which acts as the install token for the widget snippet.
 * @summary Create a new store
 */
export const createStoreBodyCurrencyDefault = `\u20B9`;

export const CreateStoreBody = zod.object({
  "owner_email": zod.email(),
  "store_domain": zod.string(),
  "whatsapp_number": zod.string(),
  "sms_number": zod.string().nullish(),
  "mode": zod.enum(['personalized', 'discount']),
  "discount_amount": zod.int().nullish(),
  "currency": zod.string().default(createStoreBodyCurrencyDefault)
})

export const CreateStoreResponse = zod.object({
  "id": zod.uuid(),
  "owner_email": zod.email(),
  "store_domain": zod.string(),
  "whatsapp_number": zod.string().nullish(),
  "sms_number": zod.string().nullish(),
  "mode": zod.enum(['personalized', 'discount']),
  "discount_amount": zod.int().nullish(),
  "currency": zod.string(),
  "created_at": zod.coerce.date()
})


/**
 * Called by the widget on load to fetch whatsapp_number, mode, discount_amount, and currency instead of using hardcoded JS constants.
 * @summary Get widget config for a store
 */
export const GetStoreConfigParams = zod.object({
  "id": zod.uuid()
})

export const GetStoreConfigResponse = zod.object({
  "whatsapp_number": zod.string().nullish(),
  "sms_number": zod.string().nullish(),
  "mode": zod.enum(['personalized', 'discount']),
  "discount_amount": zod.int().nullish(),
  "currency": zod.string()
}).describe('Minimal config needed by the widget on load')


/**
 * Update store settings from the dashboard (whatsapp number, mode, discount amount, etc.)
 * @summary Update store config
 */
export const UpdateStoreParams = zod.object({
  "id": zod.uuid()
})

export const UpdateStoreBody = zod.object({
  "whatsapp_number": zod.string().optional(),
  "sms_number": zod.string().nullish(),
  "mode": zod.enum(['personalized', 'discount']).optional(),
  "discount_amount": zod.int().nullish(),
  "currency": zod.string().optional()
})

export const UpdateStoreResponse = zod.object({
  "id": zod.uuid(),
  "owner_email": zod.email(),
  "store_domain": zod.string(),
  "whatsapp_number": zod.string().nullish(),
  "sms_number": zod.string().nullish(),
  "mode": zod.enum(['personalized', 'discount']),
  "discount_amount": zod.int().nullish(),
  "currency": zod.string(),
  "created_at": zod.coerce.date()
})


/**
 * Widget posts here when a shopper clicks the WhatsApp or SMS link. Records cart snapshot, tier, and channel.
 * @summary Log a widget capture
 */
export const createCaptureBodyTierMax = 2;



export const CreateCaptureBody = zod.object({
  "store_id": zod.uuid(),
  "tier": zod.int().min(1).max(createCaptureBodyTierMax),
  "channel": zod.enum(['whatsapp', 'sms']),
  "cart_snapshot": zod.looseObject({

}).describe('Cart items and prices at the time of capture')
})

export const createCaptureResponseTierMax = 2;



export const CreateCaptureResponse = zod.object({
  "id": zod.uuid(),
  "store_id": zod.uuid(),
  "tier": zod.int().min(1).max(createCaptureResponseTierMax),
  "channel": zod.enum(['whatsapp', 'sms']),
  "cart_snapshot": zod.looseObject({

}).describe('Cart items and prices at the time of capture'),
  "created_at": zod.coerce.date()
})


/**
 * Returns all captures for a store. Pass ?format=csv to get a CSV export instead of JSON.
 * @summary List captures for a store
 */
export const GetStoreCapturesParams = zod.object({
  "id": zod.uuid()
})

export const getStoreCapturesQueryFormatDefault = `json`;

export const GetStoreCapturesQueryParams = zod.object({
  "format": zod.enum(['json', 'csv']).default(getStoreCapturesQueryFormatDefault)
})

export const getStoreCapturesResponseCapturesItemTierMax = 2;



export const GetStoreCapturesResponse = zod.object({
  "captures": zod.array(zod.object({
  "id": zod.uuid(),
  "store_id": zod.uuid(),
  "tier": zod.int().min(1).max(getStoreCapturesResponseCapturesItemTierMax),
  "channel": zod.enum(['whatsapp', 'sms']),
  "cart_snapshot": zod.looseObject({

}).describe('Cart items and prices at the time of capture'),
  "created_at": zod.coerce.date()
})),
  "total": zod.int()
})