# lib/api-zod/src/generated/types/loginBody.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the Cart-to-WhatsApp Widget platform
 * OpenAPI spec version: 0.1.0
 */

export interface LoginBody {
  owner_email: string;
  password: string;
}