# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the Cart-to-WhatsApp Widget platform
 * OpenAPI spec version: 0.1.0
 */

export interface SignupBody {
  owner_email: string;
  /** @minLength 8 */
  password: string;
  store_domain: string;
}