# lib/api-zod/src/generated/types/store.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the Cart-to-WhatsApp Widget platform
 * OpenAPI spec version: 0.1.0
 */
import type { StoreMode } from './storeMode';

export interface Store {
  id: string;
  owner_email: string;
  store_domain: string;
  whatsapp_number?: string | null;
  sms_number?: string | null;
  mode: StoreMode;
  discount_amount?: number | null;
  currency: string;
  created_at: Date;
}