# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the Cart-to-WhatsApp Widget platform
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  success: boolean;
}